Refactor domain tab
